+++ lib/FirmwareIPC/i2c.h
@@ -1,14 +1,13 @@
 #ifndef I2C_H
 #define I2C_H
 
-#include <cstdint>
-
 #include "command.h"
 #include "fipc.h"
+#include "pico/stdio.h"
 
 class I2C : public FIPC {
 public:
-    I2C(uint8_t sda, uint8_t scl);
+    I2C(uint sda, uint scl);
     ~I2C();
 
     Command* next();
